Refactor custom transitions, create tick() function
