+++ src/lib/transitions/index.ts
@@ -1,0 +1,7 @@
+import fade from './fade';
+
+const index = {
+  fade,
+};
+
+export default index;
